requirements.txt
python-telegram-bot
python-dotenv
loguru
aiosqlite
APScheduler
requests
PyYAML
